Implement basic QuickLook preview for SGF files.
For the initial implementation, the generator loads the Spotlight metadata
attributes for the file for which a preview is requested.  These
attributes are passed to a view controller which loads a nib that has a
simple preview which is rendered into the QL context.

Future work will need to parse the SGF instead of relying on the metadata,
in order to handle files which have not been indexed and to display a
representation of the game board.